+++ src/dev/theskidster/xjge/main/App.java
@@ -82,7 +82,7 @@ public final class App {
     public static final int ALL_VIEWPORTS     = -1;
     public static final boolean DEBUG_ALLOWED = true; //TODO change this to false before building distributions.
     public static final String DOMAIN         = "xjge";
-    public static final String VERSION        = "1.2.3";
+    public static final String VERSION        = "1.5.0";
     
     private static Viewport[] viewports = new Viewport[4];
     private static Color clearColor     = Color.BLACK;
@@ -346,7 +346,7 @@ public final class App {
                     viewport.resetCamera();
                     
                     viewport.render("camera");
-                    level.render(viewport.currCamera.position, viewport.currCamera.direction, viewport.currCamera.up);
+                    level.render(viewport.currCamera);
                     viewport.render("ui");
                 glBindFramebuffer(GL_FRAMEBUFFER, 0);